requirements.txt
Flask
line-bot-sdk
psycopg2-binary